// src/generated/server/alsaglobal/market/v1/service_server.ts
export interface ListCommodityQuotesRequest {
  symbols: string[];
}

export interface ListCommodityQuotesResponse {
  quotes: CommodityQuote[];
}

export interface CommodityQuote {
  symbol: string;
  name: string;
  display: string;
  price: number;
  change: number;
  sparkline: number[];
}
// ...
export interface ServerContext {
  request: Request;
  pathParams: Record<string, string>;
  headers: Record<string, string>;
}